Implement Role Based JWT Authentication Provide Storing created and upated user id

+++ LetsDoIt.Moody/src/LetsDoIt.Moody.Application/Category/CategoryService.cs
@@ -1,10 +1,12 @@
 ﻿using NGuard;
 using System;
+using System.Runtime.Serialization;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
 
 namespace LetsDoIt.Moody.Application.Category
 {
+    using CustomExceptions;
     using VersionHistory;
     using Persistence.Entities;
     using Persistence.Repositories.Base;
@@ -52,7 +54,7 @@ namespace LetsDoIt.Moody.Application.Category
             return result;
         }
 
-        public async Task InsertAsync(string name, int order, byte[] image)
+        public async Task InsertAsync(string name, int order, byte[] image, int userId)
         {
             _logger.LogInformation($"{nameof(InsertAsync)} executing with " +
                                    $"name={name};" +
@@ -63,7 +65,8 @@ namespace LetsDoIt.Moody.Application.Category
             {
                 Name = name,
                 Order = order,
-                Image = image
+                Image = image,
+                CreatedBy = userId
             });
 
             await _versionHistoryService.CreateNewVersionAsync();
@@ -71,7 +74,7 @@ namespace LetsDoIt.Moody.Application.Category
             _logger.LogInformation($"{nameof(InsertAsync)} executed");
         }
 
-        public async Task InsertCategoryDetailsAsync(int categoryId, int order, string image)
+        public async Task InsertCategoryDetailsAsync(int categoryId, int order, string image, int userId)
         {
             _logger.LogInformation($"{nameof(InsertCategoryDetailsAsync)} executing with " +
                                    $"categoryId={categoryId};" +
@@ -82,7 +85,8 @@ namespace LetsDoIt.Moody.Application.Category
             {
                 CategoryId = categoryId,
                 Order = order,
-                Image = Convert.FromBase64String(image)
+                Image = Convert.FromBase64String(image),
+                CreatedBy = userId
             });
 
             await _versionHistoryService.CreateNewVersionAsync();
@@ -90,7 +94,7 @@ namespace LetsDoIt.Moody.Application.Category
             _logger.LogInformation($"{nameof(InsertCategoryDetailsAsync)} executed");
         }
 
-        public async Task UpdateAsync(int id, string name, int order, byte[] image)
+        public async Task UpdateAsync(int id, string name, int order, byte[] image, int userId)
         {
             _logger.LogInformation($"{nameof(UpdateAsync)} executing with " +
                                    $"id={id};" +
@@ -98,11 +102,16 @@ namespace LetsDoIt.Moody.Application.Category
                                    $"order={order};" +
                                    "image");
 
-            var entity = await _categoryRepository.SingleOrDefaultAsync(c => c.Id == id && !c.IsDeleted);
+            var entity = await _categoryRepository.GetAsync(c => c.Id == id && !c.IsDeleted);
+            if (entity == null)
+            {
+                throw new ObjectNotFoundException("Category", id);
+            }
 
             entity.Name = name;
             entity.Order = order;
             entity.Image = image;
+            entity.ModifiedBy = userId;
 
             await _categoryRepository.UpdateAsync(entity);
 
@@ -111,17 +120,22 @@ namespace LetsDoIt.Moody.Application.Category
             _logger.LogInformation($"{nameof(UpdateAsync)} executed");
         }
 
-        public async Task UpdateCategoryDetailsAsync(int id, int order, byte[] image)
+        public async Task UpdateCategoryDetailsAsync(int id, int order, byte[] image, int userId)
         {
             _logger.LogInformation($"{nameof(UpdateCategoryDetailsAsync)} executing with " +
                                    $"id={id};" +
                                    $"order={order};" +
                                    "image");
 
-            var entity = await _categoryDetailsRepository.SingleOrDefaultAsync(detail => detail.Id == id && !detail.IsDeleted);
+            var entity = await _categoryDetailsRepository.GetAsync(detail => detail.Id == id && !detail.IsDeleted);
+            if (entity == null)
+            {
+                throw new ObjectNotFoundException("Category Detail", id);
+            }
 
             entity.Order = order;
             entity.Image = image;
+            entity.ModifiedBy = userId;
 
             await _categoryDetailsRepository.UpdateAsync(entity);
 
@@ -130,11 +144,17 @@ namespace LetsDoIt.Moody.Application.Category
             _logger.LogInformation($"{nameof(UpdateCategoryDetailsAsync)} executed");
         }
 
-        public async Task DeleteAsync(int id)
+        public async Task DeleteAsync(int categoryId, int userId)
         {
-            _logger.LogInformation($"{nameof(DeleteAsync)} executing with id={id}");
+            _logger.LogInformation($"{nameof(DeleteAsync)} executing with id={categoryId}");
 
-            var category = await _categoryRepository.SingleOrDefaultAsync(c => c.Id == id && !c.IsDeleted);
+            var category = await _categoryRepository.GetAsync(c => c.Id == categoryId && !c.IsDeleted);
+            if (category == null)
+            {
+                throw new ObjectNotFoundException("Category", categoryId);
+            }
+
+            category.ModifiedBy = userId;
 
             await _categoryRepository.DeleteAsync(category);
 
@@ -143,11 +163,17 @@ namespace LetsDoIt.Moody.Application.Category
             _logger.LogInformation($"{nameof(DeleteAsync)} executed");
         }
 
-        public async Task DeleteCategoryDetailsAsync(int id)
+        public async Task DeleteCategoryDetailsAsync(int id, int userId)
         {
             _logger.LogInformation($"{nameof(DeleteCategoryDetailsAsync)} executing with id={id}");
 
-            var entity = await _categoryDetailsRepository.SingleOrDefaultAsync(detail => detail.Id == id && !detail.IsDeleted);
+            var entity = await _categoryDetailsRepository.GetAsync(detail => detail.Id == id && !detail.IsDeleted);
+            if (entity == null)
+            {
+                throw new ObjectNotFoundException("Category Detail", id);
+            }
+
+            entity.ModifiedBy = userId;
 
             await _categoryDetailsRepository.DeleteAsync(entity);
